requirements.txt
RPi.GPIO>=0.7.1
smbus2>=0.4.1
adafruit-circuitpython-mlx90614
mpu6050-raspberrypi>=1.2
adafruit-dht>=1.4.0
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
Pillow>=9.5.0
pytest>=8.0.0
